Extract OMIS validators to separate library
Extracting the validators will make it easier to maintain and test and
use across other parts of the sub-app if needed.
